extract util from cursor
